Fixed inconditionnal destructor call
The ~Optional destructor was unconditionnaly calling the destructor of the underlying
type. In the case where the Optional wasn't initialized (placement new operator wasn't called),
this could lead to invalid destructor call especially in the case where memory allocation occured.

The destructor is now called only if the Optional contains Some value.