Making sure that the sizeof Optional<T> now equals the sizeof T
Prior to that commit, this assertion might not always be true.
To dermine if the Optional was empty or not, a union with a pointer was used. If
the pointer was null, then the Optional was considered. However, it had the consequence
to change the size of the full Optional. For example, on x86_64, sizeof(void *) == 8.
For Optional<uint8_t>, this meant that sizeof (Optional<uint8_t) was also 8.

This commit replaces the pointer by a simple uint8_t. Also added some static asserts
to make sure that the sizes of the underlying object and the Optional are always equal.